requirements.txt
python-geoip-geolite2
toml
web